+++ backend/routes/payroll-calendar/index.js
@@ -3,13 +3,13 @@ const router = express.Router();
 import {
   list,
   create,
-  // update,
+  update,
 } from "../../controller/payroll-calendar.controller.js";
 import TokenValidate from "../../middleware/tokenvalidate.js";
 
 router.get("/list", list);
 router.post("/create", create);
-// router.post("/update", update);
-// router.post("/:id/update", update);
+router.post("/update", update);
+router.post("/:id/update", update);
 
 export default router;